Run e2e tests in all 3 auth methods
