+++ e2e/test/scenarios/admin/settings/sso/jwt.cy.spec.js
@@ -6,6 +6,11 @@ import {
   popover,
 } from "e2e/support/helpers";
 
+import {
+  crudGroupMappingsWidget,
+  checkGroupConsistencyAfterDeletingMappings,
+} from "./group-mappings-widget";
+
 describeEE("scenarios > admin > settings > SSO > JWT", () => {
   beforeEach(() => {
     restore();
@@ -62,6 +67,21 @@ describeEE("scenarios > admin > settings > SSO > JWT", () => {
     cy.wait("@updateSettings");
 
     cy.findByText("Success").should("exist");
+  });
+
+  describe("Group Mappings Widget", () => {
+    beforeEach(() => {
+      cy.intercept("GET", "/api/setting").as("getSettings");
+      cy.intercept("GET", "/api/session/properties").as("getSessionProperties");
+    });
+
+    it("should allow deleting mappings along with deleting, or clearing users of, mapped groups", () => {
+      crudGroupMappingsWidget("jwt");
+    });
+
+    it("should allow deleting mappings with groups, while keeping remaining mappings consistent with their undeleted groups", () => {
+      checkGroupConsistencyAfterDeletingMappings("jwt");
+    });
   });
 });
